Add MessagePayload
and parse messages in ChannelPayload. watcherCount should be optional since it's not always returned
